third_party/WebKit/LayoutTests/http/tests/credentialmanager/credentialscontainer-notifysignedin-basics.html

<!DOCTYPE html>
<title>Credential Manager: notifySignedIn() basics.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
function stubResolverChecker(c) {
    assert_equals(c, undefined, "FIXME: We're currently always returning 'undefined'.");
    this.done();
}

function stubRejectionChecker(reason) {
    assert_unreached("notifySignedIn() should not reject, but did: " + reason.name);
}

var local = new LocalCredential('id', 'name', 'https://example.com/avatar.png', 'pencil');
var federated = new FederatedCredential('id', 'name', 'https://example.com/avatar.png', 'https://federation.net');

async_test(function () {
    navigator.credentials.notifySignedIn(local).then(
        this.step_func(stubResolverChecker.bind(this)),
        this.step_func(stubRejectionChecker.bind(this)));
}, "Verify the basics of notifySignedIn(): LocalCredential.");

async_test(function () {
    navigator.credentials.notifySignedIn(federated).then(
        this.step_func(stubResolverChecker.bind(this)),
        this.step_func(stubRejectionChecker.bind(this)));
}, "Verify the basics of notifySignedIn(): FederatedCredential.");
</script>
